move InstalledFilter to metrics
