+++ core/lib/zksync_core/src/api_server/web3/namespaces/eth.rs
@@ -1,6 +1,3 @@
-use chrono::NaiveDateTime;
-use std::time::Instant;
-use vise::GaugeGuard;
 use zksync_types::{
     api::{
         BlockId, BlockNumber, GetLogsFilter, Transaction, TransactionId, TransactionReceipt,
@@ -20,7 +17,6 @@ use zksync_web3_decl::{
     types::{Address, Block, Filter, FilterChanges, Log, U64},
 };
 
-use crate::api_server::web3::metrics::{FilterType, FILTER_METRICS};
 use crate::api_server::web3::TypedFilter;
 use crate::{
     api_server::{
@@ -48,31 +44,6 @@ impl<G> Clone for EthNamespace<G> {
         Self {
             state: self.state.clone(),
         }
-    }
-}
-
-pub struct InstalledFilter {
-    pub filter: TypedFilter,
-    guard: GaugeGuard,
-    start: Instant,
-    pub last_request: NaiveDateTime,
-}
-
-impl InstalledFilter {
-    pub fn new(filter: TypedFilter, guard: GaugeGuard) -> Self {
-        Self {
-            filter,
-            guard,
-            start: Instant::now(),
-            last_request: chrono::Utc::now().naive_utc(),
-        }
-    }
-}
-
-impl Drop for InstalledFilter {
-    fn drop(&mut self) {
-        FILTER_METRICS.filter_lifetime[&FilterType::from(self.filter.clone())]
-            .observe(self.start.elapsed());
     }
 }
